pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "compliance-nlp"
version = "1.0.0"
description = "Knowledge-Graph-Augmented RAG for Multi-Framework Regulatory Gap Detection"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10,<3.12"
keywords = [
    "regulatory-compliance",
    "knowledge-graph",
    "rag",
    "legal-nlp",
    "obligation-extraction",
    "financial-regulation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial",
]

dependencies = [
    "torch>=2.1.0,<2.4",
    "transformers>=4.36.0,<4.46",
    "peft>=0.7.0",
    "datasets>=2.16.0",
    "accelerate>=0.25.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
    "neo4j>=5.14.0",
    "rank-bm25>=0.2.2",
    "pydantic>=2.5.0",
    "omegaconf>=2.3.0",
    "numpy>=1.24.0,<2.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
serve = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "prometheus-client>=0.19.0",
    "httpx>=0.25.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
train = [
    "wandb>=0.16.0",
    "deepspeed>=0.12.0",
    "bitsandbytes>=0.41.0",
]

[project.scripts]
compliancenlp-serve = "compliance_nlp.serving.server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU",
]
